1.1-ANN_spam_classifier/pyproject.toml
[project]
name = "1-1-ann-spam-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.0",
    "nltk>=3.9.2",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "ruff>=0.14.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["spam_classifier"]

[tool.uv]
package = true

[project.scripts]
spam-classifier = "spam_classifier.cli:cli"
data-clean = "spam_classifier.cli.data:clean_sms_data"
data-inspect = "spam_classifier.cli.data:inspect_labels"
data-preprocess = "spam_classifier.cli.data:data_preprocessor"
model-train = "spam_classifier.cli.model:train_classifier"
model-evaluate = "spam_classifier.cli.model:sms_spam_classifier_evaluator"
model-analyze-errors = "spam_classifier.cli.model:analyze_misclassifications"
compare-naive-bayes = "spam_classifier.cli.compare:naive_bayes_classifier"
